Got rid of ctof and ftoc functions in favor of macros (...
Author: dustin
Date: 2002/01/29 19:19:04

Got rid of ctof and ftoc functions in favor of macros (since they're
really basic, and I don't really use them, anyway).

Made an external parseSerial which asserts the length of the input
serial number is twice the length of a normal serial number for the
conversion.  Added the same assertion to h2b, just in case it's used
anywhere else, ever.

sample_devices should try to resend on ENOBUFS error, with a bit of a
pause to allow some space to free up.

The collection stuff should just link in all of mlan.a...I'm starting
to reuse more of that.

Preparing for device-specific multicast data parsing.

imported from: dustin@spy.net--projects/ibutton--fromcvs--0.1--patch-146